requirements.txt
# Core dependencies
gpt4all>=1.0.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.22.0
pandas>=1.5.0

# Web interface
flask>=2.0.0
flask-socketio>=5.3.0

# Voice interface
pyttsx3>=2.90
SpeechRecognition>=3.8.1
sounddevice>=0.4.5
pyaudio>=0.2.13

# Data handling
datasets>=2.10.0
huggingface-hub>=0.14.0

# Database
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0
alembic>=1.9.0

# CLI tools
typer>=0.9.0
typing-extensions>=4.5.0

# Utilities
tqdm>=4.65.0
requests>=2.28.0
python-dotenv>=1.0.0
schedule>=1.2.0  # For task scheduling

# Optional dependencies
# tensorflow>=2.12.0  # Uncomment if using TensorFlow
# flax>=0.6.0  # Uncomment if using Flax 